Merge remote branch 'origin/master'
Had to resolve a couple of conflicts. In the case of sst_orcmd.c, I basically took my changes as they resolved the workarounds that had been created there.

For orcmd.c, there is a header file included there (slm.h) which does NOT exist anywhere in the repo. So I commented it out.

Conflicts:
	orcm/mca/sst/orcmd/sst_orcmd.c
	orcm/tools/orcmd/orcmd.c