IPMI Plugin
Modified how the sensor labels were compared with the user fed sensor
list.

Added another new MCA parameter through which user can pass a
sensor group name, which is compared with the sensor label, and if a
substring is found, then the sensor is sampled.
for ex.: if user inputs the tag 'Temp' then all sensor with their label
containing the term Temp will be sampled.