IPMI Plugin & nodepower plugin
The IPMI plugin tries to read the bmc credentials in an endless loop in
case it fails to read for some reason, and no other compute node tries
to send the bmc credential data. Fixes #35

The nodepower plugin uses the ipmi_cmdraw command to retrieve data from
the bmc. It needs to pass the length of the response buffer to the API
so that the library knows how much memory is allocated to it and can
pack the data accordingly. The current implementation does pass the
repose length. but it doesn't initialize the length, which could lead to
non deterministic results. Initialized teh lenght to 1024 to reflect the
size of the data buffer. Fixes #33

If the ipmi_cmdraw call in the nodepower plugin fails for some reason,
then the failure is currently being ignored and the responseData is
copied out and passed to the application. Added the code to handle a
return failure of the ipmi_cmdraw call.