Fixing the way number of sockets is collected
There is an offchance that an addition object can reside in between
machine and socket levels in hwloc's topology. This could throw off the
count of number of sockets. with the new change, orcm dives to the level
of sockets and counts the number of siblings to get an accurate count.