Removed the call to orcm_db.commit() from heartbeat as it is no longer needed
The data being redirected to the analytic framework to handle how to store the data.
The heartbeat sensor no longer sends the data directly to the database, so it does
not need to call the orcm_db.commit() to persist the data.  That is now the job of
the framework that handles the data.

Change-Id: I78c05808fc5de3f8cbc6ea0d0709838e4db7fe91
Signed-off-by: Hong Quach <hong.t.quach@intel.com>