Refactor for creation of query filters in octl.
There was some repeated code for the creation of orcm filters,
this was moved to a separated function to reduce lines of code.
Also the same situation happened in error reporting, another
function was added too. Additionally, the filter list created
was free in query_db function, a separate function was added to
give explicit clean this list.

Some minor style issues were modified too.

Change-Id: I6a5f5df17e3de8e8771de3a0c9362a15c93480be
Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>