Adding query event option to octl
A new suboption for octl query has been created to retrieve all
the events stored in the DB for a given range of date-hour. The
purpose of this change is to enable a query to get all event_id
fields to be used in future queries.

The events information is retrieved from event and event_data
tables using a event_view which also has been added in this
commit.

Change-Id: Ia6a72dcefca643bbae20037fd03c43f0c619ab0a
Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>