Change event_view to include event description
The event_view was excluding the description of an event. Due to
this the view has been changed. An additional function was created
to get the event message for a given event id.
Also the header of this query and the help usage message has beed
updated.

A Klocwork issue was fixed as well.

Change-Id: I167f10ade1b12498b0b2def3def67665601c64c4
Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>