Revert use of wordexp to get argv array on interactive CLI.
The interactive CLI was affected by the wordexp function due to
this POSIX complain function was expanding the * character. With
wordexp the user would need to specify "*" instead of * on inter-
active CLI. From shell execution octl users still need to specify
the "*" to avoid command expansion.

Also with this change the quoted parameters are not supported, we
need to figure it out another way to have two different command
entries, the interactive CLI and the shell execution.

Change-Id: Ia46868cb38a5cf2ccbf2d1c2bf13d16ea9995e51
Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>