Created output_handler for message display in OCTL
Created to ensure consistent message display on OCTL, removing unnecessary
dependencies and rigidity, allowing to format the outputs and the streams.

This will remove the usage of orte_show_help API that was sending the
errors message to stdout instead of stderr.
Also will replace the macro of ORCM_UTIL_MSG that is causing warnings
due the usage of varidic macro.

Change-Id: I442349079ec25f669c85464cf21b35a4fc63e9c1
Signed-off-by: Mariana Gonzalez Velarde <mariana.gonzalez.velarde@intel.com>