SNMP Parser Refactor.
- [SENSOR] Changing configuration files for SNMP to use new XML syntax

A number of configuration files used in the test content for SNMP
has been migrated to the new XML syntax in order to have appropriate
test content for the new refactor.

- [SENSOR] Update of test content for SNMP config file refactor

Removed some test cases from snmp_parser_tests.cpp, and renaming of
a few others, in order to match new syntax of configuration files.

- [SENSOR] Use pugi parser to parse snmp xml configuration files.

- [UTIL] Moved string related functions from snmp_parser and
  snmp_collector code to a new file called string_utils.h
  at orcm/util directory.

Change-Id: Ia4c17a8f3c414d836e62389fcf3c5c95f5b989b1
Signed-off-by: Minia Oseguera <minia.oseguera@intel.com>
Signed-off-by: David Romero Antequera <david.romero.antequera@intel.com>