Fix for ormc-info version option
Fixed orcm-info to actually return all component's version when "all"
is passed as first parameter of --version option.

Updated orcm-info documentation to include "all" as a valid word for
the second parameter of --version option. It already has the same
behaviour of "full" but it was not documented.

Removed trailing white spaces.

Change-Id: I41aae12df351391dd7dbca42e370593d14a12171
Signed-off-by: Minia Oseguera <minia.oseguera@intel.com>