Fix resusage open/close test
This tests launches the entire sensor framework providing a
list of disabled sensors. This caused failures with new sensors
added that needs extra configuration to be run.

More particularly, udsensors was causing this failure because
the needed path to load sensors doesn't exists in the conditions
that resusage tests uses. However, this udsensors failure is
something that users won't see.

The solution is to remove the list of disabled sensors so everything
but resusage is disabled.

Change-Id: Icf11bd5833fdfd4d3bb91f558132676352c7d7ff
Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>